# src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Rest com cadastro, login, autenticação e CRUD",
    "version": "1.0.0",
    "description": "Essa API Rest tem como funcionalidades o cadastro, login e autenticação de usuários através do JWT (JSON Web Token), assim também como CRUD de projetos e tarefas vinculadas a esses projetos.",
    "contact": {
      "email": "lucas98fernando@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [{
      "url": "http://localhost:3000/",
      "description": "API para testes e desenvolvimento"
    },
    {
      "url": "https://www.urldeproducao.com.br",
      "description": "API de produção"
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "sumary": "Cadastro de usuários",
        "description": "Essa rota será responsável por realizar o cadastro de usuários.",
        "tags": ["Register"],
        "security": [{
          "bearerAuth": []
        }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Register"
              },
              "examples": {
                "register": {
                  "value": {
                    "name": "Lucas Fernando",
                    "email": "lucasdev@gmail.com",
                    "password": "123456"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "User already exists!"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Register"
                }
              }
            }
          }
        }
      }
    },
    "/auth/authenticate": {
      "post": {
        "sumary": "Autenticação de usuário já cadastrado",
        "description": "Essa rota será responsável por realizar a autenticação do usuário através do e-mail e senha.",
        "tags": ["Login"],
        "security": [{
          "bearerAuth": []
        }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              },
              "examples": {
                "login": {
                  "value": {
                    "email": "lucasdev@gmail.com",
                    "password": "123456"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "User not found!"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Login"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Register": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "passsword": {
            "type": "string"
          }
        }
      },
      "Login": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "passsword": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}